Wire welcome tour into the live first-run flow
- New /welcome route renders the self-contained WelcomeTourScreen with the
  real inviter name (from the user's invitation). Nav chrome hidden there.
- Onboarding routes to /welcome after areas-of-knowledge (first-Five
  generation already kicks off in the background) instead of /?welcome=1.
- "Play Now" → /daily; "I'll explore more first" → /?firstrun=1, and the home
  page shows a one-time "your first five are ready" note for that param.
- WelcomeTourScreen self-suppresses once seen (refresh/back to /welcome leaves
  for the explore target); dev replay (forced) ignores the flag.
- Removed the old overlay-on-real-home WelcomeTour component (superseded by the
  self-contained screen); repointed the dev reset button's import.

Client-side activation only — no migration; reversible.
